+++ custom_components/shinobi/configuration/helpers/const.py
@@ -22,7 +22,7 @@ DOMAIN = "shinobi"
 
 DEFAULT_PORT = 8080
 
-CONFIGURATION_MANAGER = f"cm_{DOMAIN}"
+MAIN_VIEW = f"main_view_{DOMAIN}"
 
 DATA_KEYS = [
     CONF_HOST,